requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
alphashape==1.3.1
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
astroid==3.0.1
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
atomicwrites==1.4.1
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.33.9
botocore==1.33.9
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
cloud-files==4.21.1
cloud-volume==8.28.0
cloudpickle==3.0.0
cmap==0.1.3
comm==0.2.0
compressed-segmentation==2.2.1
compresso==3.2.1
ConfigArgParse==1.7
connection-pool==0.0.3
contourpy==1.2.0
crackle-codec==0.7.0
crc32c==2.3.post0
cycler==0.12.1
datrie==0.8.2
debugpy==1.8.0
decorator==5.1.1
deflate==0.4.0
defusedxml==0.7.1
dill==0.3.7
docutils==0.20.1
dpath==2.1.6
DracoPy==1.3.0
dvidtools==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastcluster==1.2.6
fasteners==0.19
fastjsonschema==2.19.0
fastremap==1.14.0
fonttools==4.46.0
fpzip==1.2.2
fqdn==1.5.1
freetype-py==2.4.0
fuzzywuzzy==0.18.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-apitools==0.5.32
google-auth==2.25.1
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
h5py==3.10.0
hsluv==5.0.4
httplib2==0.22.0
humanfriendly==10.0
idna==3.6
igraph==0.11.3
imageio==2.33.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.27.1
ipympl==0.9.3
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-rfb==0.4.2
jupyter-ydoc==0.3.4
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==4.0.9
jupyterlab-execute-time==3.1.0
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupytext==1.16.0
k3d==2.16.0
kaleido==0.2.1
kiwisolver==1.4.5
kneed==0.8.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
molesq==0.4.0
morphops==0.1.13
msgpack==1.0.7
multiprocess==0.70.15
mypy-extensions==1.0.0
navis @ git+https://github.com/navis-org/navis.git@cf1c3a41b9708bca9c8242f1b73442113b0d26ed
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
ncollpyde==0.19.0
nest-asyncio==1.5.8
networkx==3.2.1
neuprint-python==0.4.26
neuroglancer==2.39.2
notebook==7.0.6
notebook_shim==0.2.3
nptyping==2.5.0
numpy==1.26.2
oauth2client==4.1.3
openpyxl==3.1.2
orjson==3.9.10
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
Pint==0.22
plac==1.4.1
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.3.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
Pygments==2.17.2
pylint==3.0.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pynrrd==1.0.0
pyparsing==3.1.1
pypng==0.20220715.0
pyrsistent==0.20.0
pysimdjson==5.0.2
PySocks==1.7.1
pyspng-seunglab==1.1.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-igraph==0.11.3
python-json-logger==2.0.7
python-jsonschema-objects==0.5.1
pytz==2023.3.post1
pyvis==0.3.2
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
rdata==0.10.0
referencing==0.31.1
requests==2.31.0
reretry==0.11.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
Rtree==1.1.0
s3transfer==0.8.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
selenium==4.16.0
Send2Trash==1.8.2
shapely==2.0.2
simplejpeg==1.7.2
six==1.16.0
skeletor==1.2.3
smart-open==6.4.0
smmap==5.0.1
snakemake==7.32.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
stopit==1.1.2
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
testpath==0.6.0
texttable==1.7.0
threadpoolctl==3.2.0
throttler==1.2.2
tifffile==2023.9.26
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toposort==1.10
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
traittypes==0.2.1
trimesh==4.0.5
trio==0.23.1
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
vispy==0.14.1
wcwidth==0.2.12
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.0.3
xarray==2023.11.0
y-py==0.6.2
ypy-websocket==0.8.4
yte==1.5.1
zfpc==0.1.2
zfpy==1.0.0
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
